Fix detail
fix config

Fix detail
