+++ RPi-Pico/src/bench_main.c
@@ -1,6 +1,6 @@
 /* bench_main.c
  *
- * Copyright (C) 2006-2022 wolfSSL Inc.
+ * Copyright (C) 2006-2024 wolfSSL Inc.
  *
  * This file is part of wolfSSL.
  *
