+++ RPi-Pico/src/wolfTcp.c
@@ -1,6 +1,6 @@
 /* tcp.c
  *
- * Copyright (C) 2006-2023 wolfSSL Inc.
+ * Copyright (C) 2006-2024 wolfSSL Inc.
  *
  * This file is part of wolfSSL.
  *
